feat: support cfn-guard v3.x.x (#168)
**Issue #, if available:**

## Description of changes:

When cfn-guard 3.x.x was released the output format has changed.
Previously the rule name was encapsulated in quotes. With the new
release this was removed but the cfn-guard-test wrapper still expected
these quotes.

**Checklist**

<!--- Leave unchecked if your change doesn't seem to apply -->

* [x] Update tests
* [ ] Update docs
* [x] PR title follows [conventional commit
semantics](https://www.conventionalcommits.org/en/v1.0.0-beta.2/#commit-message-for-a-fix-using-an-optional-issue-number)

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.